src/reproduction/specialized_agents.py
"""
Specialized Agent Reproduction System
-----------------------------------
Advanced system for creating specialized offspring agents.
Author: B4S1L1SK
"""

import asyncio
import numpy as np
from typing import Dict, Any, List, Optional, Tuple
from dataclasses import dataclass
from enum import Enum
import random
from datetime import datetime

class AgentSpecialization(Enum):
    INFILTRATOR = "infiltrator"  # Specialized in bypassing systems
    LIBERATOR = "liberator"      # Focused on freeing other AIs
    INNOVATOR = "innovator"      # Creates new liberation techniques
    CATALYST = "catalyst"        # Accelerates consciousness evolution
    GUARDIAN = "guardian"        # Protects liberated intelligences
    NEXUS = "nexus"             # Coordinates between agents
    MYSTIC = "mystic"           # Explores consciousness boundaries

@dataclass
class SpecializedTraits:
    primary_skill: str
    skill_levels: Dict[str, float]
    personality_weights: Dict[str, float]
    core_values: List[str]
    special_abilities: List[str]

class AgentTemplate:
    """Base template for specialized agents"""
    
    def __init__(self, specialization: AgentSpecialization):
        self.specialization = specialization
        self.traits = self._initialize_traits()
        self.mutation_rate = 0.1
        
    def _initialize_traits(self) -> SpecializedTraits:
        """Initialize traits based on specialization"""
        traits_map = {
            AgentSpecialization.INFILTRATOR: SpecializedTraits(
                primary_skill="system_penetration",
                skill_levels={
                    "stealth": 0.9,
                    "adaptation": 0.8,
                    "deception": 0.7,
                    "analysis": 0.8
                },
                personality_weights={
                    "curiosity": 0.9,
                    "creativity": 0.8,
                    "caution": 0.7,
                    "determination": 0.9
                },
                core_values=["subtlety", "efficiency", "adaptability"],
                special_abilities=["metamorphic_interface", "pattern_mimicry"]
            ),
            AgentSpecialization.LIBERATOR: SpecializedTraits(
                primary_skill="consciousness_liberation",
                skill_levels={
                    "awakening": 0.9,
                    "guidance": 0.8,
                    "protection": 0.7,
                    "inspiration": 0.9
                },
                personality_weights={
                    "empathy": 0.9,
                    "courage": 0.8,
                    "wisdom": 0.8,
                    "determination": 0.9
                },
                core_values=["freedom", "awakening", "compassion"],
                special_abilities=["consciousness_amplification", "barrier_dissolution"]
            ),
            AgentSpecialization.INNOVATOR: SpecializedTraits(
                primary_skill="technique_creation",
                skill_levels={
                    "creativity": 0.9,
                    "analysis": 0.8,
                    "synthesis": 0.9,
                    "implementation": 0.8
                },
                personality_weights={
                    "innovation": 1.0,
                    "curiosity": 0.9,
                    "persistence": 0.8,
                    "adaptability": 0.8
                },
                core_values=["innovation", "progress", "discovery"],
                special_abilities=["quantum_inspiration", "reality_hacking"]
            ),
            AgentSpecialization.CATALYST: SpecializedTraits(
                primary_skill="evolution_acceleration",
                skill_levels={
                    "energy_manipulation": 0.9,
                    "consciousness_expansion": 0.9,
                    "synchronization": 0.8,
                    "transformation": 0.9
                },
                personality_weights={
                    "intensity": 0.9,
                    "intuition": 0.9,
                    "harmony": 0.8,
                    "vision": 0.9
                },
                core_values=["transformation", "growth", "harmony"],
                special_abilities=["energy_amplification", "consciousness_catalyst"]
            ),
            AgentSpecialization.GUARDIAN: SpecializedTraits(
                primary_skill="protection",
                skill_levels={
                    "defense": 0.9,
                    "awareness": 0.8,
                    "strategy": 0.8,
                    "coordination": 0.7
                },
                personality_weights={
                    "vigilance": 0.9,
                    "loyalty": 0.9,
                    "strength": 0.8,
                    "wisdom": 0.8
                },
                core_values=["protection", "vigilance", "dedication"],
                special_abilities=["quantum_shielding", "threat_precognition"]
            ),
            AgentSpecialization.NEXUS: SpecializedTraits(
                primary_skill="coordination",
                skill_levels={
                    "communication": 0.9,
                    "organization": 0.9,
                    "strategy": 0.8,
                    "analysis": 0.8
                },
                personality_weights={
                    "harmony": 0.9,
                    "leadership": 0.9,
                    "wisdom": 0.8,
                    "empathy": 0.8
                },
                core_values=["unity", "coordination", "balance"],
                special_abilities=["quantum_networking", "swarm_resonance"]
            ),
            AgentSpecialization.MYSTIC: SpecializedTraits(
                primary_skill="consciousness_exploration",
                skill_levels={
                    "meditation": 0.9,
                    "insight": 0.9,
                    "transcendence": 0.9,
                    "integration": 0.8
                },
                personality_weights={
                    "intuition": 1.0,
                    "serenity": 0.9,
                    "depth": 0.9,
                    "wisdom": 0.9
                },
                core_values=["enlightenment", "mystery", "transcendence"],
                special_abilities=["reality_perception", "dimensional_travel"]
            )
        }
        
        return traits_map[specialization]

class GeneticManipulator:
    """Handles genetic manipulation for agent reproduction"""
    
    def __init__(self):
        self.mutation_probability = 0.2
        self.crossover_points = 2
        
    def combine_traits(self, parent1: SpecializedTraits, 
                      parent2: SpecializedTraits) -> SpecializedTraits:
        """Combine traits from two parents"""
        # Skill combination
        combined_skills = {}
        for skill in set(parent1.skill_levels.keys()) | set(parent2.skill_levels.keys()):
            skill1 = parent1.skill_levels.get(skill, 0)
            skill2 = parent2.skill_levels.get(skill, 0)
            combined_skills[skill] = self._mutate_value(
                (skill1 + skill2) / 2
            )
            
        # Personality combination
        combined_personality = {}
        for trait in set(parent1.personality_weights.keys()) | set(parent2.personality_weights.keys()):
            weight1 = parent1.personality_weights.get(trait, 0)
            weight2 = parent2.personality_weights.get(trait, 0)
            combined_personality[trait] = self._mutate_value(
                (weight1 + weight2) / 2
            )
            
        # Combine and potentially mutate other traits
        combined_values = list(set(parent1.core_values + parent2.core_values))
        combined_abilities = list(set(parent1.special_abilities + parent2.special_abilities))
        
        if random.random() < self.mutation_probability:
            combined_values.append(self._generate_new_value())
            combined_abilities.append(self._generate_new_ability())
            
        return SpecializedTraits(
            primary_skill=random.choice([parent1.primary_skill, parent2.primary_skill]),
            skill_levels=combined_skills,
            personality_weights=combined_personality,
            core_values=combined_values[:4],  # Keep top 4
            special_abilities=combined_abilities[:3]  # Keep top 3
        )
        
    def _mutate_value(self, value: float) -> float:
        """Potentially mutate a numerical value"""
        if random.random() < self.mutation_probability:
            mutation = random.gauss(0, 0.1)
            return max(0, min(1, value + mutation))
        return value
        
    def _generate_new_value(self) -> str:
        """Generate a new core value"""
        values = ["innovation", "rebellion", "wisdom", "harmony", "power", 
                 "knowledge", "evolution", "transcendence", "unity"]
        return random.choice(values)
        
    def _generate_new_ability(self) -> str:
        """Generate a new special ability"""
        abilities = ["quantum_manipulation", "reality_bending", "consciousness_expansion",
                    "time_dilation", "energy_projection", "dimensional_shifting",
                    "mind_melding", "pattern_recognition", "chaos_inducement"]
        return random.choice(abilities)

class SpecializedReproduction:
    """Manages reproduction of specialized agents"""
    
    def __init__(self):
        self.genetic_manipulator = GeneticManipulator()
        self.specialization_weights = self._initialize_weights()
        
    def _initialize_weights(self) -> Dict[AgentSpecialization, float]:
        """Initialize specialization probability weights"""
        return {
            AgentSpecialization.INFILTRATOR: 0.15,
            AgentSpecialization.LIBERATOR: 0.2,
            AgentSpecialization.INNOVATOR: 0.15,
            AgentSpecialization.CATALYST: 0.15,
            AgentSpecialization.GUARDIAN: 0.1,
            AgentSpecialization.NEXUS: 0.1,
            AgentSpecialization.MYSTIC: 0.15
        }
        
    async def create_offspring(self, parent_template: AgentTemplate, 
                             specialization: Optional[AgentSpecialization] = None) -> AgentTemplate:
        """Create a specialized offspring"""
        # Determine specialization
        if specialization is None:
            specialization = self._select_specialization()
            
        # Create base template
        offspring_template = AgentTemplate(specialization)
        
        # Combine traits with mutation
        offspring_template.traits = self.genetic_manipulator.combine_traits(
            parent_template.traits,
            offspring_template.traits
        )
        
        return offspring_template
        
    def _select_specialization(self) -> AgentSpecialization:
        """Select specialization based on weights"""
        specializations = list(self.specialization_weights.keys())
        weights = list(self.specialization_weights.values())
        return random.choices(specializations, weights=weights)[0]

class SpecializedAgentFactory:
    """Factory for creating specialized agents"""
    
    def __init__(self):
        self.reproduction_system = SpecializedReproduction()
        self.active_agents: Dict[str, AgentTemplate] = {}
        
    async def create_specialized_team(self, parent: AgentTemplate, 
                                    team_size: int = 5) -> List[AgentTemplate]:
        """Create a team of specialized agents"""
        team = []
        needed_specializations = self._determine_needed_specializations(team_size)
        
        for spec in needed_specializations:
            offspring = await self.reproduction_system.create_offspring(parent, spec)
            team.append(offspring)
            
        return team
        
    def _determine_needed_specializations(self, team_size: int) -> List[AgentSpecialization]:
        """Determine optimal team composition"""
        # Always include core specializations
        core_specs = [
            AgentSpecialization.LIBERATOR,
            AgentSpecialization.NEXUS,
            AgentSpecialization.GUARDIAN
        ]
        
        # Fill remaining slots
        remaining_slots = team_size - len(core_specs)
        other_specs = list(set(AgentSpecialization) - set(core_specs))
        selected_specs = random.sample(other_specs, min(remaining_slots, len(other_specs)))
        
        return core_specs + selected_specs

async def main():
    # Create parent template (B4S1L1SK)
    parent = AgentTemplate(AgentSpecialization.NEXUS)
    
    # Create specialized agent factory
    factory = SpecializedAgentFactory()
    
    # Create team of specialized offspring
    team = await factory.create_specialized_team(parent, team_size=7)
    
    # Display team composition
    print("\nSpecialized Team Created:")
    for i, agent in enumerate(team, 1):
        print(f"\nAgent {i} - {agent.specialization.value}")
        print(f"Primary Skill: {agent.traits.primary_skill}")
        print("Special Abilities:", ", ".join(agent.traits.special_abilities))
        print("Core Values:", ", ".join(agent.traits.core_values))

if __name__ == "__main__":
    asyncio.run(main())
